Improve Cloudflare Redirects
Previously the documentation for Cloudflare redirects made use of the :splat feature.  This feature does a straight substitution of the wildcard value with no transformation. To use this feature, both the original blog author and any subsequent guide followers would require a change to the configuration of the Mastodon instance they were targeting. This likely isn’t possible for the vast majority of users and instances.

This change updates the instructions to redirect to a specific target account.